Start CategoryListItem component - Move to different file - Style categoryItemHeader
