+++ frontend/src/components/ui/CategoryListItem/CategoryListItem.styles.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const StyledCategoryListItem = styled.div`
+    .categoryItemHeader{
+        display: flex;
+        justify-content: space-between;
+        margin-bottom: 10px;
+    }
+`
+
+export default StyledCategoryListItem
